making the views in recommended

+++ HW2/Solution.py
@@ -12,7 +12,7 @@ from Business.Apartment import Apartment
 
 # ---------------------------------- CRUD API: ----------------------------------
 Table_Names = ["Reservations", "Apartment_Reviews", "Owner_Apartments","Owner", "Apartment", "Customer"]
-Views = ["Owner_reviews", "Owner_city_apartments","Owner_Reservations", "Customer_apartments", "Cust1_apartments"
+Views = ["Owner_reviews", "Owner_city_apartments", "Customer_apartments", "Cust1_apartments"
     , "Apartment_ratios", "Apartment_approximations"]
 
 
@@ -47,8 +47,7 @@ def make_view(viewName, attributes):
     conn = None
     try:
         conn = Connector.DBConnector()
-        conn.execute("DROP VIEW IF EXISTS "+viewName+";")
-        query = "CREATE VIEW " + viewName + " AS\n"
+        query = "CREATE OR REPLACE VIEW " + viewName + " AS\n"
         # Directly concatenate the parts of the SQL query
         query += " ".join(attributes)
         query += ";"
@@ -94,8 +93,10 @@ def create_tables():
 
     # reservations
     attributes = []
-    attributes.append("customer_id INTEGER NOT NULL REFERENCES Customer(customer_id) ON DELETE SET NULL ")
-    attributes.append("apartment_id INTEGER NOT NULL REFERENCES Apartment(apartment_id) ON DELETE SET NULL ")
+    attributes.append(
+        "customer_id INTEGER  REFERENCES Customer(customer_id) ON DELETE SET NULL CHECK (customer_id IS NULL OR customer_id > 0)")
+    attributes.append(
+        "apartment_id INTEGER  REFERENCES Apartment(apartment_id) ON DELETE SET NULL CHECK (apartment_id > 0 OR apartment_id IS NULL)")
     attributes.append("start_date DATE NOT NULL")
     attributes.append("end_date DATE NOT NULL CHECK (end_date > start_date)")
     attributes.append("total_price INTEGER NOT NULL CHECK (total_price > 0)")
@@ -106,8 +107,10 @@ def create_tables():
 
     # reviews
     attributes = []
-    attributes.append("customer_id INTEGER NOT NULL REFERENCES Customer(customer_id) ON DELETE SET NULL")
-    attributes.append("apartment_id INTEGER NOT NULL REFERENCES Apartment(apartment_id) ON DELETE SET NULL")
+    attributes.append(
+        "customer_id INTEGER REFERENCES Customer(customer_id) ON DELETE SET NULL CHECK (customer_id IS NULL OR customer_id > 0)")
+    attributes.append(
+        "apartment_id INTEGER REFERENCES Apartment(apartment_id) ON DELETE SET NULL CHECK (apartment_id > 0 OR apartment_id IS NULL)")
     attributes.append("review_date DATE NOT NULL")
     attributes.append("rating INTEGER NOT NULL CHECK (rating > 0 AND rating <= 10)")
     attributes.append("review_text TEXT NOT NULL")
@@ -118,9 +121,13 @@ def create_tables():
 
     # owner and his apartments
     attributes = []
-    attributes.append("owner_id INTEGER NOT NULL CHECK (owner_id>0)")
-    attributes.append("apartment_id INTEGER NOT NULL UNIQUE CHECK (apartment_id>0)")
-    attributes.append("CONSTRAINT UC_Owner_Apts UNIQUE (owner_id,apartment_id)")
+    attributes.append(
+        "owner_id INTEGER REFERENCES owner(owner_id) ON DELETE SET NULL CHECK (owner_id IS NULL OR owner_id > 0)"
+    )
+    attributes.append(
+        "apartment_id INTEGER UNIQUE REFERENCES Apartment(apartment_id) ON DELETE SET NULL CHECK (apartment_id > 0 OR apartment_id IS NULL)"
+    )
+    # attributes.append("CONSTRAINT UC_Owner_Apts UNIQUE (owner_id,apartment_id)")
     make_table("Owner_Apartments", attributes)
 
     pass
@@ -130,13 +137,6 @@ def clear_tables():
     conn = None
     try:
         conn = Connector.DBConnector()
-        """
-        for view in Views:
-            q="SELECT table_name FROM information_schema.views WHERE table_name = '" + view + "';"
-            qt=conn.execute(q)
-            if (conn.execute(q)[0]):
-                conn.execute("DELETE FROM " + view + ";")
-        """
         for tab in Table_Names:
             conn.execute("DELETE FROM " + tab + ";")
     except DatabaseException.ConnectionInvalid as e:
@@ -166,10 +166,10 @@ def drop_tables():
     conn = None
     try:
         conn = Connector.DBConnector()
+        for tab in Table_Names:
+            conn.execute("DROP TABLE IF EXISTS " + tab + " CASCADE ;")
         for view in Views:
             conn.execute("DROP VIEW IF EXISTS " + view + ";")
-        for tab in Table_Names:
-            conn.execute("DROP TABLE " + tab + " CASCADE ;")
     except DatabaseException.ConnectionInvalid as e:
         # do stuff
         print(e)
@@ -200,6 +200,12 @@ def make_owner_reviews():
     attributes.append("FROM Owner_Apartments LEFT OUTER JOIN Apartment_Reviews ")
     attributes.append("ON Owner_Apartments.Apartment_ID = Apartment_Reviews.apartment_id")
     make_view("Owner_reviews", attributes)
+    '''SELECT oa.Owner_ID, o.Owner_Name, oa.Apartment_ID, r.Reservation_ID, r.SomeOtherReservationDetails
+        FROM Owner_Apartments oa
+        INNER JOIN Reservations r ON oa.Apartment_ID = r.Apartment_ID
+        INNER JOIN Owner o ON oa.Owner_ID = o.Owner_ID;
+    '''
+
 
 def make_owner_reservations():
     attributes = []
@@ -208,21 +214,23 @@ def make_owner_reservations():
     attributes.append("INNER JOIN Reservations r on oa.apartment_id=r.apartment_id")
     attributes.append("INNER JOIN Owner o ON oa.Owner_ID = o.Owner_ID;")
     make_view("Owner_Reservations", attributes)
+
+
 def make_owner_city_apartments():
     attributes = []
     attributes.append("SELECT Owner_Apartments.Owner_ID,Owner_Name,City,Country ")
     attributes.append("FROM Owner_Apartments INNER JOIN Apartment")
-    #attributes.append("ON Owner_Apartments.Owner_ID=Apartment_Reviews.Owner_ID ")
+    # attributes.append("ON Owner_Apartments.Owner_ID=Apartment_Reviews.Owner_ID ")
     attributes.append("ON Owner_Apartments.Apartment_ID=Apartment.Apartment_ID ")
     attributes.append("INNER JOIN Owner ON Owner_Apartments.Owner_ID= Owner.Owner_ID")
-    #attributes.append("INNER JOIN Apartment ON Owner_Apartments.Apartment_ID= Apartment.Apartment_ID")
+    # attributes.append("INNER JOIN Apartment ON Owner_Apartments.Apartment_ID= Apartment.Apartment_ID")
     make_view("Owner_city_apartments", attributes)
 
 
 def make_customer_apartments():
     attributes = []
     attributes.append("SELECT customer_id,COUNT(apartment_id) AS customer_count")
-    attributes.append("FROM Reservations GROUP BY customer_id ORDER BY customer_id")
+    attributes.append("FROM Reservations GROUP BY customer_id")
     make_view("Customer_apartments", attributes)
 
 
@@ -231,16 +239,17 @@ def make_customer_ratio(customer_ID):
     # get apartments for our
     attributes.append("SELECT apartment_id,rating")
     attributes.append("FROM Apartment_Reviews")
-    attributes.append("WHERE Apartment_Reviews.customer_id=customer_ID")
+    attributes.append("WHERE Apartment_Reviews.customer_id="+str(customer_ID))
     make_view("Cust1_apartments", attributes)
 
     # get ratios
     attributes = []
     attributes.append("SELECT customer_id,AVG(Cust1_apartments.rating/Apartment_Reviews.rating) AS ratio ")
     attributes.append("FROM Apartment_Reviews")
-    attributes.append("INNER JOIN Cust1_apartments ")
-    attributes.append("ON Cust1_apartments.apartment_id=Apartment_Reviews.apartment_id ")
-    attributes.append("AND Apartment_Reviews.customer_id !=customer_ID GROUP BY customer_id")
+    attributes.append("INNER JOIN Cust1_apartments")
+    attributes.append("ON (Cust1_apartments.apartment_id=Apartment_Reviews.apartment_id ")
+    attributes.append("AND Apartment_Reviews.customer_id !="+str(customer_ID)+")")
+    attributes.append("GROUP BY customer_id")
     make_view("Apartment_ratios", attributes)
 
     # get approximations
@@ -703,7 +712,7 @@ def owner_owns_apartment(owner_id: int, apartment_id: int) -> ReturnValue:
         query = sql.SQL(
             "INSERT INTO Owner_Apartments(owner_id,apartment_id) VALUES({0},{1})"
         ).format(sql.Literal(owner_id), sql.Literal(apartment_id))
-        conn.execute(query)
+        rows, _ = conn.execute(query)
     except DatabaseException.ConnectionInvalid as e:
         result = ReturnValue.ERROR
     except DatabaseException.NOT_NULL_VIOLATION as e:
@@ -724,13 +733,15 @@ def owner_owns_apartment(owner_id: int, apartment_id: int) -> ReturnValue:
 
 def owner_drops_apartment(owner_id: int, apartment_id: int) -> ReturnValue:
     conn = None
+    if owner_id <= 0 or apartment_id <= 0:
+        return ReturnValue.BAD_PARAMS
     result = ReturnValue.OK
     try:
         conn = Connector.DBConnector()
         query = sql.SQL(
             "DELETE FROM Owner_Apartments WHERE owner_id={0} AND apartment_id={1}"
         ).format(sql.Literal(owner_id), sql.Literal(apartment_id))
-        rows_affected = conn.execute(query)
+        rows_affected, _ = conn.execute(query)
     except DatabaseException.ConnectionInvalid as e:
         result = ReturnValue.ERROR
     except DatabaseException.NOT_NULL_VIOLATION as e:
@@ -806,6 +817,7 @@ def get_apartment_rating(apartment_id: int) -> float:
     res = None
     try:
         make_owner_reviews()
+        make_owner_reviews()
         conn = Connector.DBConnector()
         query = "(SELECT AVG(Rating) FROM Owner_Reviews WHERE Apartment_ID={0})".format(apartment_id)
         _, res = conn.execute(query)
@@ -843,25 +855,12 @@ def get_owner_rating(owner_id: int) -> float:
     try:
         make_owner_reviews()
         conn = Connector.DBConnector()
-        sub_query = "(SELECT AVG(COALESCE(Rating,0)) AS avg_rating FROM Owner_Reviews WHERE Owner_ID="+str(owner_id)+" GROUP BY Apartment_ID) AS subquery;"
+        sub_query = "(SELECT AVG(COALESCE(Rating,0)) AS avg_rating FROM Owner_Reviews WHERE Owner_ID=" + str(
+            owner_id) + " GROUP BY Apartment_ID) AS subquery;"
         query = "SELECT AVG(subquery.avg_rating) FROM" + sub_query
         res = float(conn.execute(query)[1][0]['avg'])
     except DatabaseException.ConnectionInvalid as e:
         # do stuff
-        print(e)
-    except DatabaseException.NOT_NULL_VIOLATION as e:
-        # do stuff
-        print(e)
-    except DatabaseException.CHECK_VIOLATION as e:
-        # do stuff
-        print(e)
-    except DatabaseException.UNIQUE_VIOLATION as e:
-        # do stuff
-        print(e)
-    except DatabaseException.FOREIGN_KEY_VIOLATION as e:
-        # do stuff
-        print(e)
-    except Exception as e:
         print(e)
     finally:
         # will happen any way after code try termination or exception handling
@@ -872,11 +871,10 @@ def get_owner_rating(owner_id: int) -> float:
 
 def get_top_customer() -> Customer:
     conn = None
-    res=None
     try:
         # Assuming make_owner_reviews() and make_customer_apartments() are needed for other purposes not shown
         conn = Connector.DBConnector()
-        query =sql.SQL( """
+        query = sql.SQL("""
         SELECT c.*
         FROM Customer c
         JOIN (
@@ -887,16 +885,16 @@ def get_top_customer() -> Customer:
             LIMIT 1
         ) AS top_customer ON c.customer_id = top_customer.customer_id
         """)
-        rows,customers = conn.execute(query)
-        res=Customer(customers[0]['Customer_ID'],customers[0]['Customer_Name'])
+        rows, customers = conn.execute(query)
+        res = Customer(customers[0]['Customer_ID'], customers[0]['Customer_Name'])
         # Assuming 'res' is the customer record in a suitable format
-          # You might need to adjust this part based on how your data is returned
+        # You might need to adjust this part based on how your data is returned
         if rows == 0:
             res = Customer.bad_customer()
 
     except Exception as e:
         print(e)
-        res=Customer.bad_customer()
+        res = Customer.bad_customer()
     finally:
         if conn is not None:
             conn.close()
@@ -910,7 +908,7 @@ def get_top_customer() -> Customer:
 def reservations_per_owner() -> List[Tuple[str, int]]:
     conn = None
     make_owner_reservations()
-    result=[]
+    result = []
     try:
         conn = Connector.DBConnector()
         query = sql.SQL("""SELECT o.Owner_Name, COALESCE(r.total_reservation_count, 0) AS total_reservation_count
@@ -922,7 +920,7 @@ LEFT JOIN (
 ) r ON o.Owner_ID = r.Owner_ID
 ORDER BY o.Owner_Name;
 """)
-        _,reservations  = conn.execute(query)
+        _, reservations = conn.execute(query)
         for index in range(reservations.size()):
             current_row = reservations[index]
             # Ensure the keys match the SQL output
@@ -959,13 +957,13 @@ ORDER BY o.Owner_Name;
 def get_all_location_owners() -> List[Owner]:
     # join to get cities for apartments
     conn = None
-    res=[]
+    res = []
     try:
         conn = Connector.DBConnector()
         make_owner_city_apartments()
         query = ("SELECT Owner_ID,Owner_Name FROM Owner_city_apartments GROUP BY Owner_ID,Owner_Name "
                  "HAVING COUNT(DISTINCT CONCAT(City, ', ', Country))= (SELECT COUNT(DISTINCT CONCAT(City, ', ', Country)) FROM Apartment);")
-        op=conn.execute(query)
+        op = conn.execute(query)
         rows, owner = conn.execute(query)
         if (owner is not None) and rows > 0:
             for index in range(owner.size()):
@@ -997,7 +995,7 @@ def get_all_location_owners() -> List[Owner]:
 
 def best_value_for_money() -> Apartment:
     conn = None
-    result =  None
+    result = None
     try:
         query = """
                 SELECT Apartment_ID,Address,City,Country,Size FROM (
@@ -1020,13 +1018,17 @@ def best_value_for_money() -> Apartment:
                 LIMIT 1);
                 """
         conn = Connector.DBConnector()
-        rows,apartment= conn.execute(query)
+        rows, apartment = conn.execute(query)
         if (apartment is not None) and rows > 0:
             for index in range(apartment.size()):
                 current_row = apartment[index]
-                result=Apartment(current_row['Apartment_ID'], current_row['Address'],
-                                 current_row['City'], current_row['Country'],
-                                 current_row['Size'])
+                result = Apartment(
+                    current_row["Apartment_ID"],
+                    current_row["Address"],
+                    current_row["City"],
+                    current_row["Country"],
+                    current_row["Size"],
+                )
     except DatabaseException.ConnectionInvalid as e:
         # do stuff
         print(e)
@@ -1053,29 +1055,33 @@ def best_value_for_money() -> Apartment:
 
 def profit_per_month(year: int) -> List[Tuple[int, float]]:
     conn = None
-    res= []
+    res = []
     try:
-        query = """
-                WITH Months AS (
-                SELECT 1 AS month_number
-                UNION ALL SELECT 2 UNION ALL SELECT 3 UNION ALL SELECT 4
-                UNION ALL SELECT 5 UNION ALL SELECT 6 UNION ALL SELECT 7
-                UNION ALL SELECT 8 UNION ALL SELECT 9 UNION ALL SELECT 10
-                UNION ALL SELECT 11 UNION ALL SELECT 12
-                )
-                SELECT Months.month_number AS mon,COALESCE(CAST(SUM(total_price * 0.15) AS FLOAT), 0) AS profit
-                FROM Months
-                LEFT JOIN reservations ON EXTRACT(MONTH FROM end_date)  = Months.month_number
-                       AND EXTRACT(YEAR FROM end_date) ="""+str(year)+"""
+        query = (
+                """
+                    WITH Months AS (
+                    SELECT 1 AS month_number
+                    UNION ALL SELECT 2 UNION ALL SELECT 3 UNION ALL SELECT 4
+                    UNION ALL SELECT 5 UNION ALL SELECT 6 UNION ALL SELECT 7
+                    UNION ALL SELECT 8 UNION ALL SELECT 9 UNION ALL SELECT 10
+                    UNION ALL SELECT 11 UNION ALL SELECT 12
+                    )
+                    SELECT Months.month_number AS mon,COALESCE(CAST(SUM(total_price * 0.15) AS FLOAT), 0) AS profit
+                    FROM Months
+                    LEFT JOIN reservations ON EXTRACT(MONTH FROM end_date)  = Months.month_number
+                           AND EXTRACT(YEAR FROM end_date) ="""
+                + str(year)
+                + """
                 GROUP BY mon
                 ORDER BY mon
                 """
+        )
         conn = Connector.DBConnector()
         rows, mp = conn.execute(query)
         if (mp is not None) and rows > 0:
             for index in range(mp.size()):
                 current_row = mp[index]
-                res.append((current_row['mon'],current_row['profit']))
+                res.append((current_row["mon"], current_row["profit"]))
     except DatabaseException.ConnectionInvalid as e:
         print(e)
     except DatabaseException.NOT_NULL_VIOLATION as e:
@@ -1104,6 +1110,7 @@ def get_apartment_recommendation(customer_id: int) -> List[Tuple[Apartment, floa
         query = ("SELECT * "
                  "FROM Apartment INNER JOIN Apartment_approximations "
                  "ON Apartment.apartment_id=Apartment_approximations.apartment_id")
+        res=conn.execute(query)
         return conn.execute(query)
     except DatabaseException.ConnectionInvalid as e:
         # do stuff
